Reorg docs module.
