+++ src/Docile.jl
@@ -1,10 +1,13 @@
 """
     Docile
+
+> Documentation extensions for Julia.
+
 """
 module Docile
 
 include("Utilities.jl")
 include(joinpath("Directives", "Directives.jl"))
-include("Docs.jl")
+include(joinpath("Docs", "Docs.jl"))
 
 end # module
