+++ src/Docs/Docs.jl
@@ -1,0 +1,13 @@
+"""
+    Docs
+"""
+module Docs
+
+using ..Utilities
+using ..Directives
+using Base.Meta
+
+include("hooks.jl")
+include("extensions.jl")
+
+end
